final fix (#15)
* server

* Revert "requested changes"

This reverts commit f7af90ba280855c3d6ce5900d0ff07d8965ff089.

* Boubou changes

* fix

* first five glowman

* fix server

* fix1

* Update ItemBoots.java

* Update ItemBoots.java
