+++ src/main/java/thaumicboots/api/ItemBoots.java
@@ -32,7 +32,6 @@ public class ItemBoots extends ItemArmor
         implements ITBootJumpable, ITBootSpeed, IVisDiscountGear, IRunicArmor, IRepairable {
 
     public IIcon icon;
-    public static final String GTNHLIB = "gtnhlib";
 
     public float runBonus;
     public float longrunningbonus;
@@ -45,6 +44,7 @@ public class ItemBoots extends ItemArmor
     public String iconResPath;
     public String armorResPath;
     public String unlocalisedName;
+    public EnumRarity rarity;
 
     public double jumpBonus;
 
@@ -69,6 +69,7 @@ public class ItemBoots extends ItemArmor
         iconResPath = "thaumicboots:electricVoid_16x";
         armorResPath = "thaumicboots:model/electricbootsVoidwalker.png";
         unlocalisedName = "ItemElectricVoid";
+        rarity = EnumRarity.rare; // this is less a variable, and more an indicator
     }
 
     public double getJumpModifier() {
@@ -141,8 +142,9 @@ public class ItemBoots extends ItemArmor
         return armorResPath;
     }
 
+    @Override
     public EnumRarity getRarity(final ItemStack stack) {
-        return EnumRarity.rare;
+        return rarity = EnumRarity.rare;
     }
 
     public int getRunicCharge(ItemStack arg0) {
@@ -164,7 +166,7 @@ public class ItemBoots extends ItemArmor
 
     protected float computeBonus(ItemStack itemStack, EntityPlayer player) {
         int ticks = player.inventory.armorItemInSlot(0).stackTagCompound.getInteger("runTicks");
-        float bonus = runBonus + ((ticks / 5) * longrunningbonus);
+        float bonus = runBonus + ((ticks * 0.2F) * longrunningbonus);
         return bonus;
     }
 
@@ -234,7 +236,9 @@ public class ItemBoots extends ItemArmor
         return stack != null && (stack.getItem() instanceof ItemBoots);
     }
 
+
     @Optional.Method(modid = GTNHLIB)
+    @SideOnly(Side.CLIENT) // this has to exist or it'll cause crashes
     public static void renderHUDJumpNotification() {
         Minecraft mc = Minecraft.getMinecraft();
         String text = getModeText(
@@ -244,6 +248,7 @@ public class ItemBoots extends ItemArmor
     }
 
     @Optional.Method(modid = GTNHLIB)
+    @SideOnly(Side.CLIENT) // this has to exist or it'll cause crashes
     public static void renderHUDSpeedNotification() {
         Minecraft mc = Minecraft.getMinecraft();
         String text = getModeText(
@@ -252,7 +257,7 @@ public class ItemBoots extends ItemArmor
         GTNHLib.proxy.printMessageAboveHotbar(text, 60, true, true);
     }
 
-    @Optional.Method(modid = GTNHLIB)
+    @Optional.Method(modid = "gtnhlib")
     public static String getModeText(String effect, double val) {
         String endResult = (int) val + "%";
         String result = switch ((int) val) {
