+++ src/main/java/thaumicboots/api/serverfiles/PacketJumpToggle.java
@@ -1,4 +1,4 @@
-package thaumicboots.api.network;
+package thaumicboots.api.serverfiles;
 
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
@@ -6,10 +6,12 @@ import net.minecraft.item.ItemStack;
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import io.netty.buffer.ByteBuf;
 import thaumicboots.api.ItemBoots;
 
-public class PacketSpeedToggle implements IMessage, IMessageHandler<PacketSpeedToggle, IMessage> {
+public class PacketJumpToggle implements IMessage, IMessageHandler<PacketJumpToggle, IMessage> {
 
     @Override
     public void fromBytes(ByteBuf byteBuf) {
@@ -22,13 +24,14 @@ public class PacketSpeedToggle implements IMessage, IMessageHandler<PacketSpeedT
     }
 
     @Override
-    public IMessage onMessage(PacketSpeedToggle message, MessageContext ctx) {
+    @SideOnly(Side.CLIENT)
+    public IMessage onMessage(PacketJumpToggle message, MessageContext ctx) {
         EntityPlayerMP player = ctx.getServerHandler().playerEntity;
         final ItemStack boots = ItemBoots.getBoots(player);
         if (boots != null) {
-            double jumpState = ItemBoots.changeSpeed(ItemBoots.isSpeedEnabled(boots));
-            ItemBoots.setModeSpeed(boots, jumpState);
-            PacketSpeedToggleAck ackMessage = new PacketSpeedToggleAck();
+            double jumpState = ItemBoots.changeJump(ItemBoots.isJumpEnabled(boots));
+            ItemBoots.setModeJump(boots, jumpState);
+            PacketJumpToggleAck ackMessage = new PacketJumpToggleAck();
             ackMessage.state = jumpState;
             PacketHandler.INSTANCE.sendTo(ackMessage, player);
         }
