+++ src/main/java/thaumicboots/api/serverfiles/PacketSpeedToggle.java
@@ -1,4 +1,4 @@
-package thaumicboots.api.network;
+package thaumicboots.api.serverfiles;
 
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
@@ -6,10 +6,12 @@ import net.minecraft.item.ItemStack;
 import cpw.mods.fml.common.network.simpleimpl.IMessage;
 import cpw.mods.fml.common.network.simpleimpl.IMessageHandler;
 import cpw.mods.fml.common.network.simpleimpl.MessageContext;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import io.netty.buffer.ByteBuf;
 import thaumicboots.api.ItemBoots;
 
-public class PacketJumpToggle implements IMessage, IMessageHandler<PacketJumpToggle, IMessage> {
+public class PacketSpeedToggle implements IMessage, IMessageHandler<PacketSpeedToggle, IMessage> {
 
     @Override
     public void fromBytes(ByteBuf byteBuf) {
@@ -22,13 +24,14 @@ public class PacketJumpToggle implements IMessage, IMessageHandler<PacketJumpTog
     }
 
     @Override
-    public IMessage onMessage(PacketJumpToggle message, MessageContext ctx) {
+    @SideOnly(Side.CLIENT)
+    public IMessage onMessage(PacketSpeedToggle message, MessageContext ctx) {
         EntityPlayerMP player = ctx.getServerHandler().playerEntity;
         final ItemStack boots = ItemBoots.getBoots(player);
         if (boots != null) {
-            double jumpState = ItemBoots.changeJump(ItemBoots.isJumpEnabled(boots));
-            ItemBoots.setModeJump(boots, jumpState);
-            PacketJumpToggleAck ackMessage = new PacketJumpToggleAck();
+            double jumpState = ItemBoots.changeSpeed(ItemBoots.isSpeedEnabled(boots));
+            ItemBoots.setModeSpeed(boots, jumpState);
+            PacketSpeedToggleAck ackMessage = new PacketSpeedToggleAck();
             ackMessage.state = jumpState;
             PacketHandler.INSTANCE.sendTo(ackMessage, player);
         }
