+++ src/main/java/thaumicboots/item/boots/comet/ItemQuantumCometBoots.java
@@ -1,5 +1,7 @@
 package thaumicboots.item.boots.comet;
 
+import net.minecraft.item.EnumRarity;
+
 public class ItemQuantumCometBoots extends ItemElectricCometBoots {
 
     public ItemQuantumCometBoots(ArmorMaterial par2EnumArmorMaterial, int par3, int par4) {
@@ -12,6 +14,7 @@ public class ItemQuantumCometBoots extends ItemElectricCometBoots {
         energyPerDamage = 2_500; // 4k hits, 2x prev
         visDiscount = 5;
         transferLimit = 12_000;
+        rarity = EnumRarity.rare;
         runBonus = getEMTQuantumSpeed() + 0.220F; // quantum + comet * 4
         jumpBonus = 0.9075D; // 9.5 blocks
         minimumDistance = 100.0F;
