+++ src/main/java/thaumicboots/item/boots/voidwalker/ItemElectricVoidwalkerBoots.java
@@ -8,6 +8,8 @@ import net.minecraftforge.common.ISpecialArmor;
 import net.minecraftforge.common.MinecraftForge;
 
 import baubles.common.lib.PlayerHandler;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import ic2.api.item.ElectricItem;
 import taintedmagic.common.registry.ItemRegistry;
 import thaumcraft.api.IWarpingGear;
@@ -23,7 +25,7 @@ public class ItemElectricVoidwalkerBoots extends ItemElectricBoots implements IW
     public ItemElectricVoidwalkerBoots(final ArmorMaterial material, final int j, final int k) {
         super(material, j, k);
         setCreativeTab(TabThaumicBoots.tabThaumicBoots);
-        setUnlocalizedName(unlocalisedName); // Not dead code, displays as "item.thaumicboots.null" without this
+        setUnlocalizedName(unlocalisedName);
         setBootsData();
         MinecraftForge.EVENT_BUS.register(this);
     }
@@ -43,6 +45,7 @@ public class ItemElectricVoidwalkerBoots extends ItemElectricBoots implements IW
         iconResPath = "thaumicboots:electricVoid_16x";
         armorResPath = "thaumicboots:model/electricbootsVoidwalker.png";
         unlocalisedName = "ItemElectricVoid";
+        rarity = EnumRarity.epic;
     }
 
     @Override
@@ -50,6 +53,7 @@ public class ItemElectricVoidwalkerBoots extends ItemElectricBoots implements IW
         return 5;
     }
 
+    @SideOnly(Side.CLIENT)
     protected void particles(final World world, final EntityPlayer player) {
         final FXWispEG fx = new FXWispEG(
                 world,
@@ -108,6 +112,6 @@ public class ItemElectricVoidwalkerBoots extends ItemElectricBoots implements IW
 
     @Override
     public EnumRarity getRarity(final ItemStack stack) {
-        return EnumRarity.epic;
+        return rarity = EnumRarity.epic;
     }
 }
