+++ src/main/java/thaumicboots/main/ClientProxy.java
@@ -10,11 +10,17 @@ import cpw.mods.fml.client.registry.RenderingRegistry;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import thaumicboots.main.utils.BootKeys;
 
+@SuppressWarnings("unused")
+@SideOnly(Side.CLIENT)
 public class ClientProxy extends CommonProxy {
 
-    public void load() {}
+    public void load() {
+
+    }
 
     @Override
     public void preInit(FMLPreInitializationEvent event) {
